fix(security): unify BIGINT identifiers as Long

+++ yak-security/src/main/java/io/yak/framework/security/dao/ConfigDao.java
@@ -12,7 +12,7 @@ public interface ConfigDao {
 
   public int update(ConfigPO var1);
 
-  public int deleteById(Integer var1);
+  public int deleteById(Long var1);
 
   public IPage<ConfigPO> selectPage(ConfigQueryDTO var1);
 
@@ -22,7 +22,7 @@ public interface ConfigDao {
 
   public List<String> listDistinctGroup();
 
-  public ConfigPO getbyId(Integer var1);
+  public ConfigPO getbyId(Long var1);
 
   public ConfigPO getByGroupAndName(String var1, String var2);
 }
